# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Xác minh email bằng token một lần",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/resend-verification": {
      "post": {
        "operationId": "AuthController_resendVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Gửi lại email xác minh với response chống dò email",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/google": {
      "post": {
        "operationId": "AuthController_googleLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/request-email-change": {
      "post": {
        "operationId": "AuthController_requestEmailChange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestEmailChangeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Yêu cầu đổi email và gửi xác nhận tới email mới",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/confirm-email-change": {
      "post": {
        "operationId": "AuthController_confirmEmailChange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmEmailChangeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Xác nhận email mới bằng token một lần",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/rounds/{id}/generate": {
      "post": {
        "operationId": "BracketsController_generate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "brackets"
        ]
      }
    },
    "/api/rounds/{id}/seeds": {
      "patch": {
        "operationId": "BracketsController_updateSeeds",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSeedsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "brackets"
        ]
      }
    },
    "/api/rounds/{id}/advance": {
      "post": {
        "operationId": "BracketsController_advance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "brackets"
        ]
      }
    },
    "/api/rounds/{id}/swiss/generate-next": {
      "post": {
        "operationId": "BracketsController_generateNextSwissRound",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The next Swiss round was generated exactly once",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "roundId": "round-id",
                    "bracketRound": 2,
                    "numberOfRounds": 5,
                    "matchCount": 4,
                    "matchIds": [
                      "match-id-1",
                      "match-id-2"
                    ],
                    "matches": [],
                    "bye": null,
                    "warnings": []
                  }
                }
              }
            }
          },
          "400": {
            "description": "Wrong format, incomplete current round, round limit reached, or insufficient teams"
          },
          "401": {
            "description": "Authentication is required"
          },
          "403": {
            "description": "Tournament organizer access required"
          },
          "404": {
            "description": "Round or tournament was not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate the next Swiss pairing round",
        "tags": [
          "brackets"
        ]
      }
    },
    "/api/rounds/{id}": {
      "delete": {
        "operationId": "BracketsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "brackets"
        ]
      }
    },
    "/api/rounds/{id}/bracket": {
      "get": {
        "operationId": "BracketsController_getBracket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "brackets"
        ]
      }
    },
...
  },
  "info": {
    "title": "Esports Tournament Manager API",
    "description": "Backend API contract",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {}
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Token một lần nhận trong email xác minh"
          }
        },
        "required": [
          "token"
        ]
      },
      "ResendVerificationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "player@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
      "GoogleLoginDto": {
        "type": "object",
        "properties": {}
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {}
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {}
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {}
      },
      "RequestEmailChangeDto": {
        "type": "object",
        "properties": {
          "newEmail": {
            "type": "string",
            "example": "new-email@example.com"
          },
          "currentPassword": {
            "type": "string",
            "description": "Bắt buộc với tài khoản có mật khẩu"
          }
        },
        "required": [
          "newEmail"
        ]
      },
      "ConfirmEmailChangeDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Token một lần nhận tại email mới"
          }
        },
        "required": [
          "token"
        ]
      },
...
      "UpdateSeedsDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}